feat: implement new date and time picker modules with validation and controllers for better UI interaction
